Set DaysSinceOnsetOfSymptoms
