+++ src/Covid19Radar.Api/Services/DeviceValidationAndroidService.cs
@@ -14,6 +14,7 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
+using System.Text;
 
 namespace Covid19Radar.Api.Services
 {
@@ -37,10 +38,11 @@ namespace Covid19Radar.Api.Services
         /// <summary>
         /// Validation Android
         /// </summary>
-        /// <param name="deviceVerificationPayload">Payload of Device verification. parameter</param>
         /// <returns>True when successful.</returns>
-        public bool Validation(IAndroidDeviceVerification androidDeviceVerification, byte[] expectedNonce, DateTimeOffset requestTime, AuthorizedAppInformation app)
+        public bool Validation(IAndroidDeviceVerification androidDeviceVerification, DateTimeOffset requestTime, AuthorizedAppInformation app)
         {
+            byte[] expectedNonce = GetAndroidNonce(androidDeviceVerification);
+
             var claims = ParsePayload(androidDeviceVerification.DeviceVerificationPayload);
 
             // Validate the nonce
@@ -213,6 +215,26 @@ namespace Covid19Radar.Api.Services
             public bool CtsProfileMatch { get; }
 
             public bool BasicIntegrity { get; }
+        }
+
+        public static byte[] GetAndroidNonce(IAndroidDeviceVerification submission)
+        {
+            var cleartext = GetAndroidNonceClearText(submission);
+            var nonce = GetSha256(cleartext);
+            return nonce;
+        }
+
+        private static string GetAndroidNonceClearText(IAndroidDeviceVerification submission)
+                => string.Join("|", submission.AppPackageName, submission.KeyString, GetRegionString(submission.Regions), submission.VerificationPayload);
+
+        private static string GetRegionString(string[] regions)
+            => string.Join(",", regions.Select(r => r.ToUpperInvariant()).OrderBy(r => r));
+
+        private static byte[] GetSha256(string text)
+        {
+            using var sha = SHA256.Create();
+            var textBytes = Encoding.UTF8.GetBytes(text);
+            return sha.ComputeHash(textBytes);
         }
     }
 }
